pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "koba"
version = "0.2.2"
description = "Terminal Image Renderer"
readme = "README.md"
requires-python = ">=3.9,<4.0"
license = { file = "LICENSE" }
authors = [
    {name = "simon0302010", email = "simon0302010@gmail.com"}
]
dependencies = [
    "numpy>=2.0.0",
    "Pillow>=9.5.0",
    "click>=8.0.0",
    "matplotlib>=3.9.0",
    "scikit-image>=0.24.0",
    "tqdm>=4.67.1",
    "fonttools>=4.58.0"
]

classifiers = [
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows"
]

[project.optional-dependencies]
test = ["pytest", "pytest-mock"]

[project.urls]
Homepage = "https://github.com/simon0302010/Kommandozeilenbildanzeige"

[project.scripts]
koba = "koba.main:main"
